NetApp: Deprecate E-Series drivers
The deprecation for the E-Series iSCSI/FC drivers
begins this release (Pike) and the drivers will be removed
in Queens. We will run the E-Series CI until
stable/pike is tagged EOL; no new features will be added,
however, existing code may be meaningfully refactored to
keep the drivers ready for removal.

DocImpact

Change-Id: Ie4450529b128c12cc0108fbf7e68d02399c87942