Re-arrange v3 API reference
This applies an order to the v3 API reference documents.

The idea being how does a new user come into and use the API
and what APIs should they see first when getting started.
Naturally they need to know about versions and microversions
that are avaible. Then they might need to know about volume
types before creating a volume. Once they create a volume
and know the basics of working with volumes, they can start
to take some actions on volumes. After that it's just kind
of a random free for all of APIs and the only order applied
to the rest is sorting alphabetically by title.

Change-Id: I8956de3e6907582f7115430eb959813a13e87a49