VStorage: make logging path configurable
Logging path can now be configured in shares config file.
When it's not set, the default logging path is appended
to mount options.

Also default logging path is changed because of:
When VStorage package is installed, default logging
path /var/log/vstorage is created. So this should
be a default value for logging path. Otherwise
user is forced to create logging path manually,
because share mount will fail.

Change-Id: Iba837bdd6bbf991d91f43f61f62fffebbc1877ae
Signed-off-by: Pavel Glushchak <pglushchak@virtuozzo.com>