NFS: Fix nas_secure auto mode permissions failure
When nas_secure_file_operations/permissions is set
to auto, a failure to write the secure indicator file
will result in the driver running with the secure
option set to "true".  This causes the driver to fail,
because it then can't write any other files in
unprivileged mode, either.

In "auto" mode, this should result in the driver
running with secure mode disabled.

Closes-Bug: #1685012

Change-Id: Ib51c051ab458167c8244587e14edad2ca493a926