Do proper cleanup if connect volume fails
If during the create image from volume the connect method fails with an
unexpected exception we will not be terminating the connection or
removing the export.  We will also not remove the export if the DB fails
when initializing the connection.

This patch fixes this making sure we are properly cleaning things up.

Change-Id: Ifa20366e6197bdce014d7f922d4700b684be1c23
Closes-Bug: #1685843
(cherry picked from commit b3d9b0c8a1adf52146d15dc9f5f835f0f06bf206)
Conflicts:
	cinder/tests/unit/volume/test_connection.py
	cinder/tests/unit/volume/test_image.py
	cinder/volume/driver.py
	cinder/volume/manager.py