Allow extension of volumes with snapshots for Quobyte driver
This change lifts an earlier limitation that explicitely did
not extend a volume if snapshots of that volume exist. This
limitation was lifted on the glusterfs driver and this change
mirrors that behaviour but adds using the currently active
image instead of the base file in case snapshots exist.

Partial-bug: #1687048

Change-Id: I75d25f6829d41ee2569ca3eaff63627f7c3876e0