VMware: Stop setting dynamicType
To remove the storage policy associated with a volume in vCenter,
we use a type in vSphere API called VirtualMachineEmptyProfileSpec.
Prior to vSphere 6, this type has only an optional field
'dynamicType' inherited from its base class. Since oslo.vmware
prunes empty types, we have to set this. But this property is
now removed in vSphere 6.0+. Removing the statement to set
'dynamicType' and relying on oslo.vmware not to prune types
without any fields set.

Closes-bug: #1688373
Related-bug: #1688376
Change-Id: Id9d2d1b53ee4a6069d53e6f54704632354b96bbd