Allow extension of volumes with snapshots for SMBfs driver
This change lifts an earlier limitation that explicitely did
not extend a volume if snapshots of that volume exist.

Note that we're still rejecting resize requests for volumes backed
by VHD images when snapshots exist as the image format does not
support it (this feature being available only for VHDX images).

This change also ensures that we're resizing the latest image from
the image chain.

Co-Authored-By: Lucian Petrut <lpetrut@cloudbasesolutions.com>

Partial-bug: #1687048

Change-Id: Ib4546820f405e2986c1f3456b3b393fb6974e854