Mark Infortrend drivers as unsupported
The Infortrend CI has not reported success past our CI
required period.

hecking name: Infortrend Cinder CI
  first seen: 2017-04-09 12:36:56 (25 days, 4:32:22 old)
  last seen: 2017-05-04 14:06:56 (3:02:22 old)
  last success: 2017-04-12 15:49:54 (22 days, 1:19:24 old)
  Job infortrend-cinder-volume-iscsi 3% success out of 249 comments S=9, F=240
      last success: 2017-04-12 13:45:54 (22 days, 3:23:24 old)
  Job infortrend-cinder-volume-fc 4% success out of 249 comments S=10, F=239
      last success: 2017-04-12 15:49:54 (22 days, 1:19:24 old)
Overall success rate: 3% of 250 comments

Per Cinder's non-compliance policy, this patch marks the drivers
as unsupported and deprecated and they will be removed in the
next cycle if the issue is not corrected.

Change-Id: I87a6bc0aa1854e7ae56826c5998fb8404a042d77