Mark Synology driver as unsupported
The Synology CI has not reported success past our CI required period.

Checking name: Synology DSM CI
  first seen: 2017-04-03 16:45:46 (31 days, 0:25:58 old)
  last seen: 2017-05-04 16:56:29 (0:15:15 old)
  last success: 2017-04-06 11:03:10 (28 days, 6:08:34 old)
  Job synology-dsm-cinder-iscsi 0% success out of 250 comments S=1, F=249
      last success: 2017-04-06 11:03:10 (28 days, 6:08:34 old)
Overall success rate: 0% of 250 comments

Per Cinder's non-compliance policy, this patch marks the driver as
unsupported and deprecated and it will be removed in the next
cycle if the issue is not corrected.

Change-Id: I9f74e85c5edfa64ded54b283fd382650617abddf