Mark Tegile driver as unsupported
The Tegile CI has not reported success past our CI required period.

Checking name: Tegile Storage CI
  first seen: 2017-03-17 08:39:31 (48 days, 8:32:23 old)
  last seen: 2017-04-08 03:37:59 (26 days, 13:33:55 old)
  last success:  None
  Job TegileFCDriver-dsvm-tempest 0% success out of 250 comments S=0, F=250
Overall success rate: 0% of 250 comments

Per Cinder's non-compliance policy, this patch marks the drivers
as unsupported and deprecated and they will be removed in the
next cycle if the issue is not corrected.

Change-Id: Ida70602ed8656f6897fe2709b0205f7f6ab2c6f1