Mark X-IO driver as unsupported
The X-IO CI has not reported success past our CI required period.

Checking name: X-IO OpenStack CI
  first seen: 2017-01-25 19:05:39 (98 days, 22:07:45 old)
  last seen: 2017-03-21 19:39:06 (43 days, 21:34:18 old)
  last success: 2017-02-18 04:58:47 (75 days, 12:14:37 old)
  Job dsvm-tempest-xio-iscsi 64% success out of 245 comments S=159, F=86
  Job dsvm-tempest-xio-fc 91% success out of 245 comments S=225, F=20
Overall success rate: 78% of 250 comments

Per Cinder's non-compliance policy, this patch marks the driver
as unsupported and deprecated and it will be removed in the
next cycle if the issue is not corrected.

Change-Id: I9c5b10d8fbe8471f6eb332929c6cf9dbefe8ec91