Fix forceful deletion of Consistency Group for GPFS Driver
For a operation of deleting consistency group when it contain
one or more volumes, GPFS driver is currently returning a list
of Volume objects to cinder volume manager which is expecting
a list of dictionaries to use for updating the volumes.

With this patch, GPFS will be able to delete consistency group
forcefully too.

Change-Id: Ie56c443a2a07c86001cbb9269058aaa0ccca6e84
Closes-bug: #1685355