[Trivial] SMBFS: fix passing image format using volume types
The SMBFS driver accepted specifying the image format via volume
type extra specs. The main reason was that those images were supposed
to work with Hyper-V as well as KVM (which does not apply anymore).

A previous change broke this feature, always expecting the volume
extra specs key to be 'volume_format', without considering scopes
(which are required in this case).

At the moment, this feature may still be useful, for example opting
for vhd images for compatibility reasons. Also, we're considering
adding VHDSet format support, which may be requested through volume
type extra specs.

This change fixes this issue by checking the extra specs using the
'smbfs' scope.

Closes-Bug: #1690779

Change-Id: Iab6c7792a7611aea9db4b412c676e32d3370b13e