Windows iSCSI: use configured iSCSI portals
At the moment, the Windows iSCSI driver ignores the configured
iSCSI ip addresses, picking up one of the available ones.

This change ensures that we only use the configured addresses when
reporting iSCSI portals, also ensuring that those are in fact
available.

As a side effect, this change enables multipath support.

Closes-Bug: #1369471

Change-Id: Ibfa8de9e3f27f4c7485abb9c4764b6daa5d5d6a8