Remove unused volume_types field when creating group.
Although Cinder has stored the mapping between volume_type_id and
group_id according to 'volume_type_ids' in DB when creating group.

But there is also an unused field 'volume_types' in db code that this
field is not in group ORM class.

So this patch removes it and changes the code a little to make code
more clearly.

Change-Id: I0bd99e8b1b2b74026ec1f858d92f16ee89498754
Closes-Bug: #1640085