OVO: Add reminder to add backporting code
We have a test that checks that we are bumping the version of OVOs
whenever we change the fields by checking the hashes of the OVOs and
fails with a helpful reminder to bump the version if we haven't done so.

This patch adds a reminder to that message so we don't forget to add the
backporting code, as well as mentioning that if we think that
backporting code is not necessary then it means that the version bump is
most likely not necessary as well.

TrivialFix

Change-Id: I326bb5ae1e635cd76dafec1ed671f1c47254fcbd