OVO: Fix VolumeType create not bound to a Session
We have detected a spurious failure on unittest
QuotaIntegrationTestCase.test_too_many_volumes_of_type that presented as
a "not bound to a Session; lazy load operation" error for the "projects"
attribute of a Volume Type.

This is not an error on the tests but an error on the OVO-ORM mechanism
we have to create a VolumeType instance.

When we create a VolumeType we return created volume type as we have
sent it to the DB (not actually reading from the DB the current values)
so we don't have the "projects" field loaded in the ORM instance, which
means that SQLAlchemy will try to lazy load it when we get the project
value on the "_from_db_object".

There are 3 possible solutions to the problem:

- Read from the DB the VolumeType with the project list

- Change the way we do the db_type.get so we check against the
dictionary instead of the ORM object so we don't trigger the lazy
loading: `vars(db_type).get('projects', [])`

- On the "volume_type_create" method in the DB return the projects field
with the right values.

This patch implements the last approach which fixes the VolumeType
creation issue.

In another patch we could address other potential cases where we could
be missing the project field as well, as mentioned in the second
approach, and that would require changing our fakes to use actual ORM
instances instead of dictionaries, and probably also adapting some
cinder code and additional tests.

Change-Id: I6407c4b056922227b66f2231d1e1506cd6b7a36f
Closes-Bug: #1693090
(cherry picked from commit d4496c807279a310bd149256c467de73eb25fe34)