Add an instance-uuid check on attachment_reserve
Currently the new cinder attachment_create method only allows creating
attachments for volumes that are in available or downloading state. We
intend to allow multiple attachments when multi-attach is ready to
land and then manage everything off of attachment-id's, but for now
we're not enabling that.

In order to allow Nova to continue to do things like live-migration
which is really a special case of multi-attach we'll add an instance
uuid check that allows this flow to continue until we fully turn on
multi-attach.

Just check on the _attachment_reserve, if there's an instance_uuid
specified, and it matches the existing attachments ignore the status
check failure.

Change-Id: I507d7a552c713b097cbcee1598757f53d70f0b4d
Closes-Bug: #1694530