VMAX driver - conflicting threads effecting rollback
There is an issue where conflicting threads tried to act on the same
VMAX volume in certain rollback situations (i.e. one process returns
it to the default storage group while the other process deletes it),
leading to inconsistencies later. This patch rectifies the issue by
locking the volume at the relevant method. We are directly
submitting to the stable ocata branch because our SMI-S driver has
been replaced with REST.

Change-Id: I8d5b38c7ca4ef6609e79c9368c593bc588aa97b5
Closes-Bug: #1694544