VMAX driver - failure in initiator group cleanup
There is an issue where the initiator group does not get cleaned up
if the host name is hyphenated. This patch rectifies the issue by
getting the host name from the connector object instead of parsing it
from the masking view name. This fix is being applied directly to
stable ocata because the SMI-S driver has been replaced by REST in
Pike.

Change-Id: I064ec043c136a46c9d4f3439f849c1843f76a4cf
Closes-Bug: #1694546