Allow Pure drivers to handle detach with no host
Apparently force-detach is ok with no connector
object passed in. This changes the driver to assume
that no connector means just detach the volume from
anything its attached to.

I updated the base driver to document this quirk
too.

This should only ever happen with force detach and
"old" style connections, which means multi-attach
shouldn't be in the picture so this is kind of
safe... probably..

Change-Id: Ia0710d66bdbc85afffdb6e4f22f3ce6ca642ec75
Closes-Bug: #1696215