SMBFS: enhance volume cloning
This change enhances the volume clone operation by simply copying
the image file if the cloned volume doesn't have any snapshots.

This way, we don't have to convert a chain of images. Also, this
means that we may benefit from copy offloading.

This improvement especially helps image caching, which relies on
efficient volume cloning.

This is basically implemented in the base remotefs module, any
similar driver leveraging it may enable it by switching a flag,
assuming that it implements the required methods.

Note that the copied image needs to be extended according to the
new volume size. We cannot simply call the 'extend_volume' method
as most of the RemoteFS based drivers will apply a lock to it, in
which case we'd get into a deadlock. For this reason, we follow
the common practice among those drivers, involving a helper that
won't get a lock decorator.

Change-Id: I96ac2f019d522ebbeb6c28fc4c11b41d9ba8fdc0