Fix migration completion for New Attach API's
The old attach workflow let us do things that we probably shouldn't,
like attaching a volume without actually going through Cinder and
running the attach cmd to set the status.

With the new Attach API  we try and avoid a lot of the little special
things by having everything go through an attachment object.  A
challenge with this is the retype or migration of an in-use volume.
The old flow was to attach the new volume but not mark it as
attached until the migration completion callback.  The new API's
discourage this and set the volume to attached.. well, when you
attach it.

The cleanest way to deal with this right now is first to actually
check on the nova side if the swap was initiated by Cinder or not,
and always (when using new attach API) have Nova delete the original
attachment after the swap/copy.

On the cinder side in that case we only need to update the attachment
object, so remember attachment-2 (destination) is a new volume but we
go through the volume entry in the db and swap it's ID with the original
volume-id, so we need to do a similar swap in the attachment table.

This keeps the secondary attachment record which is valid, and just does
the crazy swappy magic of the volume_id entry. This also maintains
backward compatability for the old Attach/Detach flow.

Change-Id: I23391300468cc5e2f45a12d5ad8617ab48c21e6a
Closes-Bug: #1697060