VMAX driver - None connector object in a terminate_connection
Similar to Pure driver the VMAX driver required the connector object
not be None. See https://review.openstack.org/#/c/471861/ for more
details. This fix allows for the connector to be None by forcing a
detach regardless. This was not cherry-picked from master as the SMIS
driver has been replaced by REST in Pike

Change-Id: I9805273770e8c7ea9ab2530ca71a36c96108a835
Closes-Bug: #1697661