Set default lvm_type to auto
We've had an lvm_type=thin for quite some time, and we've wanted to
make it Cinders default for a while, but there's no clean way to
migrate on upgrades.  So, we have a compromise, detect if the
deployment has volumes (including the thin pool).  If there are no
LVs or if the thin pool exists, use thin.  Otherwise, use the old
thick lvm-default.

This patch makes the default setting "auto" instead of "default"
(which is actually thick).

DocImpact: Need to update docs to reflect the default setting for
new deployments is auto (thin)

Change-Id: If25e814eb6282fef9a1dd30d17c80eab6860e275