NetApp: block drivers fail when volume missing.
The block driver will fail when attempting to perform a Cinder delete
when the volume does not exist on the backend. The Cinder volume status
is then set to error_deleting even though the volume no longer exists.
This fix logs a warning if the volume doesn't exist on the backend and
allows Cinder to successfully remove the Cinder volume.

Change-Id: I259f72cab4e6843e74f33fb566fc6201346c9ea1
Closes-Bug: #1682928