Change Huawei driver's private data storing place
This patch is to fix issues at two circumstances:
1. Create volume from image by non-admin user, this operation
will fail because Huawei driver tries to record LUN WWN at admin
metadata, however non-admin user doesn't have the admin metadata
access right, so exception is raised.
2. Restore volume backup to 3rd volume, this operation will also
restore backup volume's metadata to the 3rd volume, but the data
recorded in metadata is strong correlative to the corresponding
LUN at backend storage, so this will break the one-to-one
relationship between the restored volume and the LUN at storage.

To fix the issues above, we moved all LUN private data to
provider_location, stored as a dict string, not use admin
metadata or metadata to store LUN private data.

Change-Id: Id18622935ecaa1715491b4ede48c7b627e43d3fa
Closes-Bug: #1699109