[XIV] Fix exception message format issue
Bunch of exceptions' messages can't be rendered due to
misuse of CinderException. CinderException only support
rendered message when initializing, otherwise the
unrendered version will directly be used as output.

NOTE: As the "_(message) % params" already used in this driver
and these exceptions are in the driver layer, this change takes
the same measure of forcing translation to fix the issue.
Also this patch doesn't intend to remove the translate symbol
in the "LOG.error" method.

Change-Id: I26900ba5c643f66265b873f9c0377b89c471584a