Replace OS_AUTH_TYPE with OS_AUTH_SYSTEM in rc
We've removed OS_AUTH_TYPE support in favor of OS_AUTH_SYSTEM
as of 2.2.0.  This change just updates the example rc file to
reflect that.  While we're at it we'll also go ahead and update
the example entry for the endpoint IP to point to the localhost.

Of course this will need to be changed if you're running the
client on a different node than the container.

Closes-Bug: #1696635

Change-Id: Ia865aab9ec6dd1fac5a74a22a56bc6accc8b60b8