VMware: Fix volume cascade delete
VMDK driver does not allow snapshot deletion if the volume
is attached. Currently we assume the volume to be attached
if it's state is not 'available', which is wrong. During
cascade delete, the volume status is set to 'deleting'.
Therefore, snapshot deletion fails even if the volume is
not attached. Fixing this by using volume attachment info
to check for attached volumes.

Change-Id: I83874d6d287fe3950b003f7eac74db9a1602432e
Closes-bug: #1701393