Fix image cache db query when limits are enabled
We were missing the context parameter... and apparently
never tested with cache limits enabled. This fixes
it and updates the unit tests to check for it too.

I'll follow up with a devstack patch to set some
limits by default too so we cover more of this
code and have it behave like a production deployment
in gate jobs.

Change-Id: Ia6d43b35843682bd201e57bb2941ca0243606412
Closes-Bug: #1701696
(cherry picked from commit 70719d028a86f79fe5a70a013c5dd6e67b202517)