VMAX driver - volume and snapshot treated as dicts in attach snapshot
Volume and snapshot objects are treated as dicts in attach and detach
snapshot functionality. This causes an exception in stable ocata when
accessing members of the class. This has not been cherry picked from
the master branch as the SMI-S driver has been replaced by REST and
fixes are no longer applicable to Pike.

Change-Id: I619d61c7d07d4845ce46640bb28f8937052c4f84
Closes-Bug: #1702065