Rename method 'delete' to 'delete_backup' in backup drivers
Once we'll get generic-backup-implementation blueprint done it will be
confusion to have 'delete' method in volume drivers. After this change
we'll have two methods: 'delete_volume' and 'delete_backup' which should
be clear for maintainers.

This patch doesn't change any public API so it should be clear for
operators and users.

Change-Id: I7f975066c1c848fec7e930a9d71b7225f3250acc
Related-blueprint: generic-backup-implementation