Fix attachment_delete for reserved case
In attachment-delete we attempt a little short cut
in cases where the volume is reserved but not actually
attached to an instance. In this case we don't have a
connector and we never actually sent anything to the
driver, so we take a short cut in cinder.volume.api
and we just call an object.destroy().

The problem with this is that the object doesn't do
the proper/needed context settings and checks and just
fails due to policy. We should clean this up and add
the extra db update steps and also include the
notification messages.

Change-Id: Ib448f0548739cd42edf03eab07f305b3ac184d8f
Closes-Bug: #1702512