Fix errors preventing doc builds
There were some syntax problems that came over with the
migration of the install docs in commit:
88be92ee719baff6d5f2f7e7e073b118db2dc2e0

It appears that the docs build for Cinder is more strict
than what was being done in openstack-manuals as the
syntax issues were causing our build to fail. The next step
in our doc migration is to fix these errors so we can do doc
builds.

Change-Id: I7282b2fa2efb50cbb6490ff36f69b123959b6dde