Migrate volume management CLI documentation
The next part of the great migration of documentation
from openstack-manuals to the Cinder repo.

This patch brings over the volume management CLI
documentation and adds it to the index for our documentation.

Note, like the other patches, this is just intended to migrate
the content and get it accessible again. Future patches will
endeavor to clean up and better organize the content.

Change-Id: I197c4e53539974ba4dff4ec441ad98823cb9d22d