Remove dothill driver options
The dothill driver has been marked unsupported for some time
now and has been disabled as a standalone driver.  Since the
driver can no longer be used it doesn't make sense to keep
the dothill_* options in the code.

The drivers that use the dothill driver use lenovo_ and hpmsa_
named options instead.

This patch removes the unused options.

Change-Id: I96afd3d7ecde564d8e79f325d3713e52d8529921