Don't verify format with autodetect after image conversion
This is not useful for raw Glance images and causes
spurious failures, because raw images could contain any
format of data that qemu-img info would detect.

This caused image->volume to fail if a Glance image was
uploaded as a raw image but contained a qcow2 header.

Futhermore, this check is not providing any real safety
because it would only ensure that when "qemu-img convert"
runs and returns zero, it does what we told it to do.

We are better off not running this extraneous check.

Closes-Bug: #1703892

Change-Id: I71f16f9513d07abd83d50a25872bdde5dc3f2cc6