VMAX driver - performance improvements
There are a couple of performance issues on attaching and detaching
volumes in the VMAX cinder driver. This patch alleviates these issues
by using the Unisphere "move" API where possible, instead of
removing the volume from one storage group and then adding it to
another. This patch also changes the frequency that we call
"rest.find_snap_vx_sessions", which makes an expensive RESTAPI call
to gather all the replication information of a volume, calling it
only when necessary.

Change-Id: I4e0028a6bb0f26f6369eb1c9677dd7ca4eae2c0c
Closes-Bug: #1704155