ibm-storage: enable FC zonning to all ports
During 'initialize_connection' we rely on the storage ability to know
which hosts can access it. This is in some cases not accurate. So we
need to enable hosts to connect to all the storage FC ports.

Change-Id: I41697632d48f7cbcbe008867671f8196b5543c94
Closes-Bug: #1702291
(cherry picked from commit 6d3d64c6bbafb57a8b34554d4baef115cc17d962)