Update group status
In the volume manager, the group status is updated before the
group_id is set/unset in the volumes during update_group.
This is wrong because there's a slight chance that the group
could be modified after it becomes 'available' but before
group_id is updated in all the volumes.

This patch fixes it by updating the group status after the
group_id in all volumes are updated.

Closes-Bug: #1704810
Change-Id: I451db1c8c49c4319323e5e0328c89c115d6047d1
(cherry picked from commit 845cbd7dc0d88556c8a5398e4dfa49c46f0fba26)