Add input format check for qos_spec update
If the input format for qos_spec is none string
value in dict, Cinder will raise 500 error.

This patch added the format check and raised 400
InvalidInput error to users. We also should
update the creation process of qos_spec, to
move the validation to qos_specs.py as Eric
suggested.

Co-Authored-By: wanghao<sxmatch1986@gmail.com>

Change-Id: I5cb68af74c68cd53bc3b84e705ca3472859c1ec3
Closes-bug: #1704988
Closes-Bug: #1710805