VMAX driver - align VMAX QOS settings with front end
Currently, the QOS feature used in VMAX doesn't work with the
frontend parameters for the total bandwidth and the total IOPs
for guest. VMAX QOS should use the total_bytes_sec and total_iops_sec
so QOS enforcement can be accomplished at both the front end
(hypervisor) and the back end (VMAX), if required.

Change-Id: I996543b57920309969ebd31262bc4c859a08fea0
Closes-Bug: #1705042