LVM: Activate LV before creating exports
In some circumstances (such as after a reboot),
especially with thin LVM, the LV may not be activated, so
creating a target from it will fail.

Activate the LV before creating a target.

When using permanent=True for activation, don't supply
the -K/ignoreskipactivation argument.  This is rejected
by LVM 2.02.171, and is not a reasonable request, since
it asks to simultaneously unset a flag, and also to ignore
the flag (which should now be unset) during the same
activation operation.

This fixes a failure to clone a volume from another thin
LVM volume in the same pool.

Closes-Bug: #1701545

Change-Id: I925ece8833cc99a95c4cf693804ca3698f8cb114