Mock execute in unit test
This patch fixes the test_volume_create_after_thin_creation test
to mock execute instead of defining an execute method inside
the test. Since the unit test is run twice with different options,
the unit test asserts different things depending on the configuration
of how the test is being run.

Change-Id: Ia841ca55b36aa19d4d6b174239f4d4200793d4e5
Closes-bug: 1676646