NetApp: Fix SVM scoped user startup exception
Due to an issue discovered in certain versions of ONTAP the driver may
not start for a SVM scoped user. When this error occurs an exception is
raised, by the driver, that includes the text "User not permitted to
query Data ONTAP volumes". The bug fix resolves this issue for a SVM
scoped user that has the correct role assigned (i.e. vsadmin).

Change-Id: Ic000ba6694eca72ae78437f7d773277fd67b0395
Closes-Bug: #1705738
(cherry picked from commit 158ccc7836f1aad0514767d53c9805e681463f4b)