Handle concurrent volume creation request
Cinder passes same 'image_id', from which the
lock_name is created. This 'lock_name' is used
by 'tooz.locking' to acquire the lock.

In this flow, based on 'lock_name' it creates the
'key' and stored in 'etcd'. etcd3gw client does
not allow to modify the same key several times
within one transactionn. so to aviod this, pass
unique 'lock_name' which in turn creates unique
key.

Change-Id: Ib78ba6aa1fcc488aa2ab7db25a37021f206519dd
Closes-Bug: #1705442