VMAX driver - Error attaching failed over volumes.
Attempting to attach a volume to a server on FC after a replication
failover fails with "error: list index out of range". This is because
'common.get_target_wwns_from_masking_view' does not check if the
volume is failed over. This patch rectifies the issue. This patch also
adds some locks around RDF groups to avoid concurrency issues.

Change-Id: Ie04513903b7e75e6373124c164007fd141acdcc6
Closes-Bug: #1707181