[DOC BLD FIX] Name files for inclusion properly
The Sphinx build was issuing hundreds of warning for
duplicate labels.  The reason for this was because
all of the tables we had were named as .rst files.
This would cause Sphinx to process them twice and it would
think there were duplicated sections.

There were two ways this could be handled:  1) Exclude the tables
directory from the docs build. 2) Name the files as they should
have always been so they weren't build twice.  Given that
option 1 just masked the problem, I am implementing this
patch using option 2.

So, all the tables we are including in our documentation
have been named from .rst to .inc and the .. include::
directives that use them have also been updated to use
the new file name.

Change-Id: If395eb652f7e3b789bcbd5e6d6d05954c23d8d8a