Ibm_storage - fix failover_replication.
running failover_replication of a group with volumes fails.
This happens because we are not switching between the backends
at the end of the process, and we are not returning the correct
update fot the volume. This patch fixes it.

Change-Id: Ic76cc4738132f993e805559cff1c40cb055cc26d
Closes-Bug: 1709100
(cherry picked from commit 6addf0ed00ad22a616cc806c90e5d1917661dbbf)