[Docs] Add documentation contribution docs
Now that we are in charge of managing our documentation I thought
we should have a section in the Contributor Guide on
how to contribute documentation.  This patch adds that.

It also resolves build WARNINGS due to the README files I added
not being included in any toctrees and adds the missing README
file to the install directory.

Change-Id: I8e12628b439a400ebd1ee6d691673a16e2c2f9d2