[DOC FIX] Create missing index pages
There was a mistake in the documentation migration.  The
landing pages in the admin and configuration sections should
have been moved to index.html.  That step didn't happen which
is causing it to appear that that content is missing from
the main OpenStack docs website.

This patch updates our main index page and moves the existing
landing pages to index.rst.

Change-Id: Idb96a09f748d3bbebf6eb768dfd2b758da545096