Change pure drivers default replication interval
With newer versions of Purity the previous default
value would not be allowed, the minimum is 60 minutes
(3600 seconds).

While this does change a default config value, it
goes from a non-working-crash-the-driver-at-startup
value to a working one.

Along with this I'm bumping the driver version
numbers and fixing typo in the docs about the purity
version (this bug was found with REST API 1.5).

Change-Id: I21148f0d56cb7ce376107253aa67fc808fa40f9f
Closes-Bug: #1709720