VMAX driver - Ocata assigning QoS attributes to the correct SG
QoS is tied to the volumetype which is tied to the storage group.
The problem arises when a bootable volume is created or a volume
is attached to an instance, the volume is moved to a different
storage group which doesn't have the QoS parameters. This has not
been cherry picked from the master branch as the SMI-S driver has
been replaced by REST and fixes are no longer applicable to Pike.

Change-Id: I4d1969851d1499deed360d851c1512d93ca5a6c2
Closes-Bug: #1712050