Allow v3.0 volume metadata API calls
Commit I575635258c10f299181b8e4cdb51a7ad1f1be764 added some optimization
to a few of the volume metadata API calls. It added a check incorrectly
for the microversion for those changes such that if you were not using
microversion 3.15 or above, it would fail the API call.

This changes the microversion check to correctly either perform the
new requested functionality or fail back to the base 3.0/2.0 API
behavior.

Change-Id: I09ad8200f73273dae8e5d926939620b71d9210e8
Closes-bug: #1712192