Fix allocated capacity report on non pool drivers
The allocated_capacity_gb is being calculated by the core Cinder code and then
added to the driver provided stats, but it only works if the driver is
reporting data of pools and it doesn't work if they are reporting the
stats as the whole backend.

When reporting stats as the whole backend, which is the case of the RBD
driver, we will not send allocated_capacity_gb to the manager.

This patch fixes this by adding the stats from the special fixed pool
that is created by the _count_allocated_capacity method for drivers that
don't report pools information on stats from drivers that don't report
pools information.

Closes-Bug: #1712549
Closes-Bug: #1706057
Change-Id: Ia7ff6f05df85b1752b4e0353a734f3071dc8ff39