[DOC BLD FIX] Fix duplicate content warnings
The contributor/drivers.rst file was using 'automodule' to
pull in docstrings from the code.  This is an efficient
way to get the documentation into the contributor part
of the documentation but it was causing Sphinx to think
that it was building the documentation for those sections
twice.

To resolve this I needed to add :noindex: to the automodule
arguments.  That fixed the WARNINGS that were being produced
and appears to produce properly formatted HTML.

Change-Id: I4107843a7363791b4607fe6d206340035e56e377