Use backup_use_temp_snapshot setting for default
Commit I0173d397760bc17c01af792b4c4cc2bcbf45fedd introduced a config
option intended for the specific driver, but added to driver.py and
worded such that would appear to be a generic setting. Rather than
going through a deprecation cycle for this and introducing a new
vendor specific option, make this setting be used as the default for
all drivers.

The default for drivers was False, and the default value for this
config option is False, so nothing changes without operator action.
Those drivers that want to force this behavior are already overriding
the call to get this base driver option, so they will not change
either. And now that the NFS driver (mostly) supports snapshots, we
should be able to start expecting that as a basic volume driver
functionality.

Change-Id: I5b193512137d845f5b1ac46c758d748a19a90739
Partial-bug: #1686403