[DOC BLD FIX] Make opts sections lowercase
The sections in the cinder/opts.py file that are currently
upper case (I.E. BACKEND) cause a sphinx warning to be
generated.  When we switch to treating warnings as errors
this will break our doc build.

This patch updates the code that generates the cinder/opts.py
file to make sure that all of the section names are lower-case
with the exception of the 'DEFAULT' section.  With this
change in place we will not getting Sphinx warnings during
a doc build from the automatic config file generation.

Change-Id: Ibabbcaf7618623219690a2b27a2e548ea00be9d1