[DOC BLD FIX] Fix remote-code-block warnings
There were a number of configuration sample files
(i.e. rootwrap.conf, policy.json and api-paste.ini)
that were trying to be included with the remote-code-block
directive which is no longer supported.

I have copied over the latest sample files for Pike and
made them .inc files.  In the future these should be dynamically
created, but at this point we just need to get something out there
to fix the Sphinx build warnings.  The work to make things
dynamically generated requires more invasive changes like
policy-in-code.

I also discovered that the link from the landing page for the
configuration page had a broken link to the sample cinder.conf
file.  I fix that problem in this patch as well.

Change-Id: I2b587abbdeaee1cfe32b100c98d99131759c2171