api-ref: Remove unused parameter defs
During the initial migration of the api-ref docs, there were
several parameter definitions pulled over that are not actually
used anywhere in the docs.

It may be that some of these _should_ be referenced, but at
least as of right now they are not. Rather than carrying them
around on the chance that they could be used down the road, we
should just clean them out for now.

Change-Id: Ic9018608b00705c80df05243647f5c81124610df