api-ref: Clarify os-host GET behavior
Getting all hosts returns every Cinder service host, regardless
of which service is running on the host. But getting a specific
host will fail if you try it for a host not running cinder-volume.

This looks to be the behavior from the beginning, but there was
nothing denoting this, causing errors and confusion for code that
thought it could get a host from the list, then get detailed info
about that host. The details return volume and snapshot counts,
so it really only makes sense for cinder-volume, so the code ends
up returning a 404 for anything else.

This API design seems a little disjointed, but since this is how
it appears to have always been, just make sure the api docs have
the right details for potential API consumers to know what to
expect.

Change-Id: If53279cfcbbde1297bb2e55e17d17b473e7d0d6e
Closes-bug: #1691144