api-ref: Normalize response codes
We had a mix of formatting for our API response codes. This
makes it so all have a leading space, no trailing comma, and
no empty Error response labels.

This also addresses a formatting issue with due to the spacing
between the Normal and Error lines that was causing the two to
run together in the formatted HTML, making it harder to read.

Change-Id: Ic411ee9f671c48ce60bda21984dafe55135685ba