VMAX driver - concurrently deleting volumes can fail
When concurrently deleting volumes, sometimes the task fails leaving
the volume in an 'error_deleting' state. This only occurs when
image_volume_cache_enabled is True because a temp snapvx session is
created. This same issue also occurs when there is a batch delete
of snapshots from the same source volume. The fix locks the delete
snapvx session using the source DeviceID as key.

Change-Id: I1f72a8567faa78fec447b9e0a6f948514b43a864
Closes-Bug: #1714922