Optimizes volume creation in the Quobyte Driver
This change takes two steps to optimize the creation
of volumes with the Quobyte driver:
1. It removes a superfluous create_volume call during creation of
a new volume from an snapshot. This creation result is not used with
Quobyte but subsequently overwritten by the following qemu-img convert
operation.
2. It changes the creation regular file based volumes. For
regular files fallocate is used instead of writing the file via dd.

Partial-Bug: #1715078

Change-Id: I8a922f4a7a9e62658469ef31eb55c6d4b9ab8477