Remove Glance v1 API support
The Glance v1 API is about to be removed. Cinder support for using the
current v2 API was added in Pike, but we allowed selection of which
version to use through the glance_api_version config option. That
option was also deprecated in Pike.

This removes the config option to allow selection and switches things
over to only use Glance API v2.

Change-Id: Ice379db9ae83420bacf9e96e242c7515930eae86