Adds a Cache for Volumes Created from Snapshots with Quobyte
This is not related to the Cinder image cache.

In order to speed up the creation of multiple volumes from a single
snapshot this change adds a cache of volumes created by merging a
snapshots backing chain to the Quobyte driver.
This behaviour can be activated via a new config option
'quobyte_volume_from_snapshot_cache'.
Instead of merging a snapshots backing chain into a new volume each time
a volume is created from this snapshot, the new implementation merges the
backing chain into a volume in the new volume cache. New volumes to be
created from that snapshot are then copied from the cached volume and
no longer require the merging process.
Merging happens only for the first time when the cached volume copy is
created from a specific snapshot. Subsequent creations of volumes from
this snapshot are simply copied from the cache which requires no costly
backing chain merge.

Partial-Bug: #1715078

Change-Id: I2142b1c0a0cc2c4f85794416e702a326d3406b9d